refacotr : 컴포넌트 위치 및 명명 수정 PostGrid 컴포넌트화 , map 돌리는 TrackImage 격자로 수정
